refactoring: breadcrumb component  - component  - docs
