+++ docs/components/breadcrumb.md
@@ -1,7 +1,7 @@
 <script setup>
-import BreadcrumbExample from './breadcrumb/examples/BreadcrumbExample.vue'
-import BreadcrumbSolidExample from './breadcrumb/examples/BreadcrumbSolidExample.vue'
-import BreadcrumbCustomIconsExample from './breadcrumb/examples/BreadcrumbCustomIconsExample.vue'
+import FwbBreadcrumbExample from './breadcrumb/examples/FwbBreadcrumbExample.vue'
+import FwbBreadcrumbExampleSolid from './breadcrumb/examples/FwbBreadcrumbExampleSolid.vue'
+import FwbBreadcrumbExampleCustomIcons from './breadcrumb/examples/FwbBreadcrumbExampleCustomIcons.vue'
 </script>
 # Vue Breadcrumb - Flowbite
 The breadcrumb component is an important part of any website or application that can be used to show the current location of a page in a hierarchical structure of pages.
@@ -10,50 +10,84 @@ Flowbite includes two styles of breadcrumb elements, one that has a transparent 
 
 ## Default breadcrumb
 
-<BreadcrumbExample />
-
+<fwb-breadcrumb-example />
 ```vue
-<script setup>
-import { Breadcrumb } from 'flowbite-vue'
-</script>
 <template>
-<Breadcrumb>
-  <BreadcrumbItem href="#" home>
-    Home
-  </BreadcrumbItem>
-  <BreadcrumbItem href="#">
-    Projects
-  </BreadcrumbItem>
-  <BreadcrumbItem>
-    Flowbite
-  </BreadcrumbItem>
-</Breadcrumb>
+  <fwb-breadcrumb>
+    <fwb-breadcrumb-item home href="#">
+      Home
+    </fwb-breadcrumb-item>
+    <fwb-breadcrumb-item href="#">
+      Projects
+    </fwb-breadcrumb-item>
+    <fwb-breadcrumb-item>
+      Flowbite
+    </fwb-breadcrumb-item>
+  </fwb-breadcrumb>
 </template>
+
+<script setup>
+import { FwbBreadcrumb, FwbBreadcrumbItem } from 'flowbite-vue'
+</script>
 ```
 
 ## Solid Breadcrumb
 
-<BreadcrumbSolidExample />
-
+<fwb-breadcrumb-example-solid />
 ```vue
-<script setup>
-import { Breadcrumb } from 'flowbite-vue'
-</script>
 <template>
-<Breadcrumb solid>
-  <BreadcrumbItem href="#" home>
-    Home
-  </BreadcrumbItem>
-  <BreadcrumbItem href="#">
-    Projects
-  </BreadcrumbItem>
-  <BreadcrumbItem>
-    Flowbite
-  </BreadcrumbItem>
-</Breadcrumb>
+  <fwb-breadcrumb solid>
+    <fwb-breadcrumb-item home href="#">
+      Home
+    </fwb-breadcrumb-item>
+    <fwb-breadcrumb-item href="#">
+      Projects
+    </fwb-breadcrumb-item>
+    <fwb-breadcrumb-item>
+      Flowbite
+    </fwb-breadcrumb-item>
+  </fwb-breadcrumb>
 </template>
+
+<script setup>
+import { FwbBreadcrumb, FwbBreadcrumbItem } from 'flowbite-vue'
+</script>
 ```
 
 ## Custom Icons
 
-<BreadcrumbCustomIconsExample />
+<fwb-breadcrumb-example-custom-icons />
+```vue
+<template>
+  <fwb-breadcrumb>
+    <fwb-breadcrumb-item home href="#">
+      <template #home-icon>
+        <svg class="w-4 h-4 mr-2" fill="none" stroke="currentColor" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg">
+          <path d="M3 7v10a2 2 0 002 2h14a2 2 0 002-2V9a2 2 0 00-2-2h-6l-2-2H5a2 2 0 00-2 2z" stroke-linecap="round" stroke-linejoin="round" stroke-width="2" />
+        </svg>
+      </template>
+      Home
+    </fwb-breadcrumb-item>
+    <fwb-breadcrumb-item href="#">
+      <template #arrow-icon>
+        <svg class="w-4 h-4 text-gray-400 mr-2" fill="none" stroke="currentColor" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg">
+          <path d="M14 5l7 7m0 0l-7 7m7-7H3" stroke-linecap="round" stroke-linejoin="round" stroke-width="2" />
+        </svg>
+      </template>
+      Projects
+    </fwb-breadcrumb-item>
+    <fwb-breadcrumb-item>
+      <template #arrow-icon>
+        <svg class="w-4 h-4 text-gray-400 mr-2" fill="none" stroke="currentColor" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg">
+          <path d="M14 5l7 7m0 0l-7 7m7-7H3" stroke-linecap="round" stroke-linejoin="round" stroke-width="2" />
+        </svg>
+      </template>
+      Flowbite
+    </fwb-breadcrumb-item>
+  </fwb-breadcrumb>
+</template>
+
+<script setup>
+import { FwbBreadcrumb, FwbBreadcrumbItem } from 'flowbite-vue'
+</script>
+```
