+++ docs/components/breadcrumb/examples/FwbBreadcrumbExample.vue
@@ -1,0 +1,22 @@
+<template>
+  <div class="vp-raw">
+    <fwb-breadcrumb>
+      <fwb-breadcrumb-item
+        home
+        href="#"
+      >
+        Home
+      </fwb-breadcrumb-item>
+      <fwb-breadcrumb-item href="#">
+        Projects
+      </fwb-breadcrumb-item>
+      <fwb-breadcrumb-item>
+        Flowbite
+      </fwb-breadcrumb-item>
+    </fwb-breadcrumb>
+  </div>
+</template>
+
+<script setup>
+import { FwbBreadcrumb, FwbBreadcrumbItem } from '../../../../src/index'
+</script>
