+++ docs/components/breadcrumb/examples/FwbBreadcrumbExampleSolid.vue
@@ -1,0 +1,22 @@
+<template>
+  <div class="vp-raw">
+    <fwb-breadcrumb solid>
+      <fwb-breadcrumb-item
+        home
+        href="#"
+      >
+        Home
+      </fwb-breadcrumb-item>
+      <fwb-breadcrumb-item href="#">
+        Projects
+      </fwb-breadcrumb-item>
+      <fwb-breadcrumb-item>
+        Flowbite
+      </fwb-breadcrumb-item>
+    </fwb-breadcrumb>
+  </div>
+</template>
+
+<script setup>
+import { FwbBreadcrumb, FwbBreadcrumbItem } from '../../../../src/index'
+</script>
