+++ src/components/FwbBreadcrumb/FwbBreadcrumb.vue
@@ -1,13 +1,17 @@
 <template>
-  <nav :class="breadcrumbWrapperClasses" aria-label="Breadcrumb">
+  <nav
+    :class="breadcrumbWrapperClasses"
+    aria-label="Breadcrumb"
+  >
     <ol :class="breadcrumbClasses">
       <slot name="default" />
     </ol>
   </nav>
 </template>
+
 <script lang="ts" setup>
 import { toRefs } from 'vue'
-import { useBreadcrumbClasses } from '@/components/Breadcrumb/composables/useBreadcrumbClasses'
+import { useBreadcrumbClasses } from './composables/useBreadcrumbClasses'
 
 const props = defineProps({
   solid: {
@@ -16,5 +20,5 @@ const props = defineProps({
   },
 })
 
-const { breadcrumbWrapperClasses, breadcrumbClasses } = useBreadcrumbClasses(toRefs(props))
+const { breadcrumbClasses, breadcrumbWrapperClasses } = useBreadcrumbClasses(toRefs(props))
 </script>
