+++ src/index.ts
@@ -2,6 +2,8 @@ export { default as FwbAlert } from './components/FwbAlert/FwbAlert.vue'
 export { default as FwbAvatar } from './components/FwbAvatar/FwbAvatar.vue'
 export { default as FwbAvatarStack } from './components/FwbAvatar/FwbAvatarStack.vue'
 export { default as FwbAvatarStackCounter } from './components/FwbAvatar/FwbAvatarStackCounter.vue'
+export { default as FwbBreadcrumb } from './components/FwbBreadcrumb/FwbBreadcrumb.vue'
+export { default as FwbBreadcrumbItem } from './components/FwbBreadcrumb/FwbBreadcrumbItem.vue'
 
 export { default as Accordion } from './components/Accordion/Accordion.vue'
 export { default as AccordionPanel } from './components/Accordion/AccordionPanel.vue'
@@ -15,8 +17,6 @@ export { default as Tab } from './components/Tabs/components/Tab/Tab.vue'
 export { default as Dropdown } from './components/Dropdown/Dropdown.vue'
 export { default as FlowbiteThemable } from './components/utils/FlowbiteThemable/FlowbiteThemable.vue'
 export { default as FlowbiteThemableChild } from './components/utils/FlowbiteThemable/components/FlowbiteThemableChild/FlowbiteThemableChild.vue'
-export { default as Breadcrumb } from './components/Breadcrumb/Breadcrumb.vue'
-export { default as BreadcrumbItem } from './components/Breadcrumb/BreadcrumbItem.vue'
 export { default as Badge } from './components/Badge/Badge.vue'
 export { default as TheCard } from './components/Card/TheCard.vue'
 export { default as Carousel } from './components/Carousel/Carousel.vue'
